+ Updated map format to correct triangle indice index and texture coord lookup index per indice slot

+++ kex3_anubis/source/renderer/renderScene.h
@@ -42,7 +42,7 @@ private:
     
     kexWorld                *world;
     kexRenderView           *view;
-    float                   clipY;
+    int                     clipY;
     int                     vertCount;
     int                     triCount;
 };
